requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2023.1
Django==4.2.18
django-crispy-forms==2.1
django-js-asset==3.0.1
django-mptt==0.15.0
Pillow==10.1.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
